Figured out I have to use AutoMapper and ViewModels
I think we need ViewModels combined with AutoMapper to map those ViewModels to
make sure we are guarded against overposting and underposting errors. One
way or another, when using AutoMapper, it's checking against the
validation on the Item class which doesn't happen if you map the
properties of the ViewModel to the Model yourself. It's a few lines of
code so that shouldn't be a problem, however I have only used AutoMapper
4.2.1 and the CreateMap function is deprecated so I might look
into that later.