Created tests for Acc. Records and Books Controllers
These tests check some basic functionality that the controllers' action
methods must provide. A lot of these tests are commented out. This is
because for migrations and what not the project needs to compile. The
proper solution to this is to use repositories and view models and mocks,
which was my original solution, however considering that the project would
balloon too quickly, it has not been implemented yet. More tests to come