Figured out how to create Acession Reg. with Item
Accession Register requires an item to already exist to be created.
This means that the Create action method and view need to be fed that Item
when it is GET'ed for the first time, i.e. a new Accession Record must
already have an item associated with it. This is done by creating a new
Model in the GET Create method and assigning it an Item from an ItemId.

Then in the view, all the required Item properties (Title and Id) need to
be sent individually as HiddenFor() fields.

This commit also moves to using "Services" which encapsulate the DbContext
as it's Repository. Each model has an IService and a concrete Service
associated with it. This is to make testing easier. Adding Services was
the original purpose of this commit but I forgot to branch.