Update the Accession Record tests.
These tests verify and test that certain methods exist and behave a
certain way. A detailed note explaining the behavior of each test is
included. Please note that these tests are not complete, nor shall they
ever be because as the project develops and changes, so will the tests.

Also, they do not include design tests, for example there's no test that
tests if the "Index" page has a link to Add Book or whatever. There's two
reasons for this, first being that testing that is fairly difficult and
secondly it may not be the absolute right thing to do, because that is
more of a front end UX decision and can change a lot.