Created IssuedItem tests
This commit adds tests for the Issue Item functionality. It adds a Model,
the Service, IService and the default, generated Controller. However, it
has been left intentionally incomplete. This is because how an item (or
rather it's copy) should be selected to be issued remains debatable. I
don't know the best way for it and I believe it should be discussed with
the team.