Update IssuedItemTests and add MemberService
This commit updates the IssuedItem tests to test a new workflow. This
workflow is outlined in the issued-item-ui.png in the Drive folder. The
changes made in this commit (theoretically) allow us to work without using
ViewModels and complicating things further. It also adds the Member
Service layer that the IssuedItemController must use.

The tests written in this commit may not be perfect. They should be
revisited for review at a later date as soon as possible.