Added Login Views and controllers
This commit begins the process of adding User Manageement to the project.
The commit modifies the LibraryContext and other important parts of the
code to make use of the Identity and OWIN packages.

This commit also includes modifications to the Web.config to use a
connectionString, which seems to solve a lot of database connection
problems some contributers were having. If you don't have a Library
database, please add one.